+++ packages/dodo-ui/src/lib/components/Form/Button/Button.stories.svelte
@@ -9,7 +9,7 @@
     componentSizeOptions,
     componentVariantOptions,
   } from '@flightlesslabs/dodo-core-attributes';
-  import Theme from '$lib/components/Layout/Theme/Theme.svelte';
+  import Theme from '../../Layout/Theme/Theme.svelte';
 
   const description = `
  A flexible Button component .