+++ packages/dodo-ui-bits/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @flightlesslabs/dodo-ui-bits
 
+## 0.1.14
+
+### Patch Changes
+
+- Updated dependencies
+  - @flightlesslabs/dodo-ui@0.32.0
+
 ## 0.1.13
 
 ### Patch Changes