+++ packages/dodo-ui-date/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @flightlesslabs/dodo-ui-date
 
+## 0.4.15
+
+### Patch Changes
+
+- Updated dependencies
+  - @flightlesslabs/dodo-ui@0.32.0
+
 ## 0.4.14
 
 ### Patch Changes